requirements.txt
# requirements
flask
numpy
PyPDF2
python-docx
tqdm
sentence-transformers
nltk
chromadb
openai
python-dotenv
